# app/Models/Servicio.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Servicio extends Model
{
    use HasFactory;

    protected $table = 'servicios';

    protected $fillable = ['nombre', 'descripcion', 'categoria'];

    // Muchos a Muchos
    public function hoteles(){
        return $this->belongsToMany(Hotel::class);
    }
}