# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion Hoteles",
        "description": "desciption",
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://gestion-hotel.test/"
        }
    ],
    "paths": {
...
        "/api/servicio": {
            "get": {
                "tags": [
                    "Servicio"
                ],
                "summary": "Obtener lista de servicios paginada",
                "operationId": "b3faaf3e5a77daf3c6ff34d7f371b64d",
                "responses": {
                    "200": {
                        "description": "Lista de servicios paginada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Servicio"
                ],
                "summary": "Crear un nuevo servicio",
                "description": "Crea un nuevo servicio con los datos proporcionados",
                "operationId": "777f1233f8a253590e41385954209741",
                "requestBody": {
                    "description": "Datos necesarios para crear un servicio",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreServicioRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Servicio creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Servicio"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/servicio/{servicio}": {
            "get": {
                "tags": [
                    "Servicio"
                ],
                "summary": "Obtener un servicio",
                "description": "Devuelve los detalles de un servicio específico.",
                "operationId": "77ab61fb84d9a3333d654e7c878d57cd",
                "parameters": [
                    {
                        "name": "servicio",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del servicio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Servicio"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Servicio"
                ],
                "summary": "Actualizar datos de un servicio",
                "description": "Actualiza los datos de un servicio específico.",
                "operationId": "b19c7abd845d502558de656fc6c07026",
                "parameters": [
                    {
                        "name": "servicio",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutServicioRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Servicio actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Servicio"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Servicio"
                ],
                "summary": "Eliminar un servicio",
                "description": "Elimina un servicio específico por ID.",
                "operationId": "a0f5c9b862158c4b56f14f0b3f28d769",
                "parameters": [
                    {
                        "name": "servicio",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "ok"
                            }
                        }
                    }
                }
            }
        },
        "/api/servicio/{servicio}/hoteles": {
            "get": {
                "tags": [
                    "Servicio"
                ],
                "summary": "Obtener los hoteles asociados a un servicio",
                "description": "Devuelve la lista de hoteles asociados a un servicio específico.",
                "operationId": "57545b32c4c31e080bba4b3c2876e8ba",
                "parameters": [
                    {
                        "name": "servicio",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de hoteles asociados al servicio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Hotel"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Servicio no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Servicio no encontrado"
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Hotel": {
                "title": "Hotel",
                "description": "Modelo de un hotel",
                "required": [
                    "nombre",
                    "direccion"
                ],
                "properties": {
                    "id": {
                        "description": "ID del hotel",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del hotel",
                        "type": "string"
                    },
                    "direccion": {
                        "description": "Dirección del hotel",
                        "type": "string"
                    },
                    "telefono": {
                        "description": "Teléfono de contacto del hotel",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del hotel",
                        "type": "string",
                        "format": "email"
                    },
                    "sitioWeb": {
                        "description": "Sitio web del hotel",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Servicio": {
                "title": "Servicio",
                "description": "Modelo de un servicio",
                "required": [
                    "nombre",
                    "descripcion",
                    "categoria"
                ],
                "properties": {
                    "id": {
                        "description": "ID del servicio",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del servicio",
                        "type": "string"
                    },
                    "descripcion": {
                        "description": "Descripción del servicio",
                        "type": "string"
                    },
                    "categoria": {
                        "description": "Categoría del servicio",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StoreServicioRequest": {
                "title": "Store Servicio Request",
                "description": "Datos requeridos para crear un nuevo servicio",
                "required": [
                    "nombre",
                    "descripcion",
                    "categoria"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del servicio",
                        "type": "string"
                    },
                    "descripcion": {
                        "description": "Descripción del servicio",
                        "type": "string"
                    },
                    "categoria": {
                        "description": "Categoría del servicio",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PutServicioRequest": {
                "title": "Put Servicio Request",
                "description": "Datos requeridos para actualizar un servicio",
                "required": [
                    "nombre",
                    "descripcion",
                    "categoria"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del servicio",
                        "type": "string"
                    },
                    "descripcion": {
                        "description": "Descripción del servicio",
                        "type": "string"
                    },
                    "categoria": {
                        "description": "Categoría del servicio",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
...
}